move CompactionWorkerBuilder to builder.rs
