+++ slatedb-cli/src/main.rs
@@ -2,7 +2,7 @@ use crate::args::{parse_args, CliArgs, CliCommands, GcResource, GcSchedule};
 use chrono::{TimeZone, Utc};
 use object_store::path::Path;
 use slatedb::admin::{self, Admin, AdminBuilder};
-use slatedb::compaction_worker::CompactionWorkerBuilder;
+use slatedb::CompactionWorkerBuilder;
 use slatedb::compactor::{
     CompactionRequest, CompactionSchedulerSupplier, SizeTieredCompactionSchedulerSupplier,
 };
